add command fallback and some config option

+++ src/main/java/com/deepseasaltyfish/externalview/networking/AttributePayload.java
@@ -7,10 +7,10 @@ import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.resources.ResourceLocation;
 import com.deepseasaltyfish.externalview.ExternalView;
 
-public record AttributePayload(int value) implements CustomPacketPayload {
+public record AttributePayload(double value) implements CustomPacketPayload {
     public static final Type<AttributePayload> TYPE = new Type<>(ResourceLocation.fromNamespaceAndPath(ExternalView.MOD_ID, "attribute"));
     public static final StreamCodec<FriendlyByteBuf, AttributePayload> CODEC = StreamCodec.composite(
-            ByteBufCodecs.INT, AttributePayload::value,
+            ByteBufCodecs.DOUBLE, AttributePayload::value,
             AttributePayload::new
     );
